Revert back changes
